bioinformatic_api_server/requirements.txt
attrs==21.4.0
biopython==1.79
black==22.1.0
click==8.0.4
flake8==4.0.1
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pyclean==2.1.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pysam==0.18.0
pytest==7.1.0
tomli==2.0.1
tornado==6.1
typing-extensions==4.1.1
